\name{agricolae-package}
\alias{agricolae-package}
\alias{agricolae}
\docType{package}
\title{
Statistical Procedures for Agricultural Research
}
\description{
This package contains functionality for the Statistical Analysis of experimental
designs applied specially for field experiments in agriculture and plant breeding.
}
\details{
\tabular{ll}{
Package: \tab agricolae\cr
Type: \tab Package\cr
Version: \tab 1.3-2\cr
Date: \tab 2020-01-18\cr
License: \tab GPL\cr
}
Planning of field experiments: lattice, factorial, RCBD, CRD, Latin Square,
Youden, Graeco, BIB, Alpha design, Cyclic, augmented block, split and strip plot
Designs. Comparison of multi-location trials: AMMI, Index AMMI Stability
(biplot, triplot), comparison between treatments: LSD, Bonferroni and
other p-adjust, HSD, Waller, Student Newman Keuls SNK, Duncan, REGW, Scheffe;
Non parametric tests: Kruskal, Friedman, Durbin, Van Der Waerden, Median.
Analysis of genetic experiments: North Carolina designs, LinexTester, 
Balanced Incomplete Block, Strip plot, Split-Plot, Partially Balanced Incomplete 
Block, analysis Mother and baby trials (see data RioChillon).
Resampling and simulation: resampling.model, simulation.model, montecarlo,
lateblight Simulator for potato. Ecology: Biodiversity Index, Path Analysis.
Soil Uniformity: Smith's Index. Cluster Analysis: Consensus Cluster.
Descriptive statistics utilities: *.freq
}
\author{
Felipe de Mendiburu
Statistical Engineer
Master in Systems Engineering
Professor of Applied Statistics

Maintainer: Felipe de Mendiburu <fmendiburu@lamolina.edu.pe>
}
\references{

{De Mendiburu, Felipe (2009). Una herramienta de analisis estadistico para la
investigacion agricola. Tesis. Universidad Nacional de Ingenieria (UNI-PERU).}

{Universidad Nacional Agraria La Molina, Lima-PERU.
Facultad de Economia y Planificacion
Departamento Academico de Estadistica e Informatica}
}
\keyword{ package }
%- Experimental Design
