\name{agricolae-package}
\alias{agricolae-package}
\alias{agricolae}
\docType{package}
\title{
Statistical Procedures for Agricultural Research
}
\description{
This package contains functionality for the Statistical Analysis of experimental
designs applied specially for field experiments in agriculture and plant breeding.
}
\details{
\tabular{ll}{
Package: \tab agricolae\cr
Type: \tab Package\cr
Version: \tab 1.0-2\cr
Date: \tab 2007-02-06\cr
License: \tab This software may be re-distributed freely and used under the GPL license 2.0 or higher\cr
}
Planning of field experiments: lattice, factorial, RCBD, CRD, Latin Square, Greaco, BIB, Alpha design. 
Comparison of multi-location trials: AMMI Stability (biplot and triplot),
Comparison between treatments: LSD, HSD, Waller, Kruskal, Friedman, Durbin, Van Der Waerden.
Resampling and simulation: resampling.model, simulation.model,
Ecology: Indices Biodiversity, path analysis
Uniformity Soil: Index Smith's
}
\author{
Felipe de Mendiburu

Maintainer: Felipe de Mendiburu <f.mendiburu@cgiar.org>
Team leader: Reinhard Simon <r.simon@cgiar.org>
}
\references{
International Potato Center (CIP)
}
\keyword{ package }
%- Experimental Design
