\name{bar.err}
\alias{bar.err}
%- bar.err.
\title{ Plotting the standard error or standard deviance of a multiple 
comparison of means}
\description{
 It plots bars of the averages of treatments and standard error or standard 
 deviance. It uses the objects generated by a procedure of comparison like LSD,
 HSD, Kruskal and Waller-Duncan. 
 }
\usage{
bar.err(x,variation=c("SE","SD","range","IQR"),horiz=FALSE, bar=TRUE,...)
}
\arguments{
  \item{x}{ object means of the comparisons the LSD.test, HSD.test,...,etc }
  \item{variation}{ SE=standard error, range=Max-Min or IQR=interquartil range}
  \item{horiz}{ Horizontal or vertical bars }
  \item{bar}{paint bar}
  \item{\dots}{ Parameters of the function barplot()}
}
\details{
  x: data frame formed by 5 columns: name of the bars, height, level
  out: LSD.test, HSD, waller.test, scheffe.test, duncan.test, SNK.test, 
  friedman, kruskal, waerden.test and Median.test. 
}
\value{
A list with numeric vectors giving the coordinates of all the bar midpoints drawn.
\item{x}{ eje-1 coordinate }
\item{height}{ eje-2 coordinate by group}
}
\author{ Felipe de Mendiburu }
\seealso{\code{\link{LSD.test}}, \code{\link{HSD.test}}, 
\code{\link{waller.test}}, \code{\link{kruskal}}, \code{\link{bar.group}} }
\examples{
library(agricolae)
data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
out <- waller.test(model,"virus", console=TRUE,
main="Yield of sweetpotato\ndealt with different virus")
oldpar<-par(mfrow=c(2,2),cex=1)
bar.err(out$means,variation="range",horiz=TRUE,xlim=c(0,45),angle=125,density=6,
 main="range")
bar.err(out$means,variation="SD",ylim=c(0,45),col=colors()[30],
 main="Standard deviation",density=8)
bar.err(out$means,variation="SE",horiz=TRUE,xlim=c(0,45),density=8,
 col="brown",main="Standard error")
bar.err(out$means,variation="range",ylim=c(0,45),bar=FALSE,col="green",
 main="range")
par(mfrow=c(1,2),cex=1)
bar.err(out$means,variation="range",ylim=c(0,45),bar=FALSE,col=0)
abline(h=0)
# horiz = TRUE
bar.err(out$means,variation="SE",horiz=TRUE,xlim=c(0,45),bar=FALSE,col=0)
#startgraph
par(oldpar)
#endgraph
}
\keyword{ aplot }% at least one, from doc/KEYWORDS

