\name{bar.group}
\alias{bar.group}
%- bar.group.
\title{ Plotting the multiple comparison of means }
\description{
 It plots bars of the averages of treatments to compare.
 It uses the objects generated by a procedure of comparison like LSD, HSD,
 Kruskall, Waller-Duncan, Friedman or Durbin. It can also display the 'average' 
 value over each bar in a bar chart.
}
\usage{
bar.group(x, horiz = FALSE, ...)
}
\arguments{
  \item{x}{ Object created by a test of comparison }
  \item{horiz}{ Horizontal or vertical bars }
  \item{\dots}{ Parameters of the function barplot() }
}
\details{
x: data frame formed by 5 columns: name of the bars, height and level of the bar.
}
\value{
A list with numeric vectors giving the coordinates of all the bar midpoints drawn.
\item{x}{ eje-1 coordinate }
\item{height}{ eje-2 coordinate by group}
}

\author{ Felipe de Meniburu }

\seealso{
\code{\link{LSD.test}}, \code{\link{HSD.test}}, \code{\link{kruskal}}
, \code{\link{friedman}}, \code{\link{durbin.test}}, \code{\link{waller.test}} 
, \code{\link{plot.group}}
}

\examples{
# Example 1
library(agricolae)
data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
comparison<- LSD.test(model,"virus",alpha=0.01,group=TRUE)
print(comparison$groups)
oldpar<-par(cex=1.5)
bar.group(comparison$groups,horiz=TRUE,density=8,col="blue",border="red", xlim=c(0,50),las=1)
title(cex.main=0.8,main="Comparison between\ntreatment means",xlab="Yield",ylab="Virus")
# Example 2
library(agricolae)
x <- 1:4
y <- c(0.29, 0.44, 0.09, 0.49)
xy <- data.frame(x,y,y)
par(oldpar)
oldpar<-par(cex=1.5)
bar.group(xy,density=30,angle=90,col="brown",border=FALSE,ylim=c(0,0.6),lwd=2,las=1)
par(oldpar)
}
\keyword{ aplot }% at least one, from doc/KEYWORDS

