% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carolina.R
\name{carolina}
\alias{carolina}
\title{North Carolina Designs I, II and III}
\usage{
carolina(model, data)
}
\arguments{
\item{model}{Constant}

\item{data}{Data frame}
}
\value{
\strong{model} model analysis (I, II or III) of caroline design

\strong{variance}  additive variance of male, female and male.female interaction.
}
\description{
Statistic analysis of the Carolina I, II and III genetic designs.

model = 1,2 and 3 is I, II and III see carolina1,2 and 3.
}
\examples{


library(agricolae)
data(DC)
carolina1 <- DC$carolina1
# str(carolina1)
output<-carolina(model=1,carolina1)
output[][-1]

carolina2 <- DC$carolina2
# str(carolina2)
majes<-subset(carolina2,carolina2[,1]==1)
majes<-majes[,c(2,5,4,3,6:8)]
output<-carolina(model=2,majes[,c(1:4,6)])
output[][-1]

carolina3 <- DC$carolina3
# str(carolina3)
output<-carolina(model=3,carolina3)
output[][-1]

}
\references{
Biometrical Methods in Quantitative Genetic Analysis, Singh, Chaudhary. 1979
}
\seealso{
\code{\link{DC}}
}
\author{
\enumerate{
         \item Felipe de Mendiburu (\email{fmendiburu@lamolina.edu.pe})
         \item Muhammad Yaseen (\email{myaseen208@gmail.com})
         }
}
\keyword{models}
