% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clay.R
\docType{data}
\name{clay}
\alias{clay}
\title{Data of Ralstonia population in clay soil}
\format{
A data frame with 69 observations on the following 3 variables.
}
\source{
Experimental field.
}
\usage{
data(clay)
}
\description{
An evaluation over a time period.
}
\details{
\itemize{
       \item{\strong{per.clay}} per.clay
       \item{\strong{days}} days
       \item{\strong{ralstonia}} ralstonia
       }
}
\examples{

library(agricolae)
data(clay)
str(clay)

}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
