% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\docType{data}
\name{corn}
\alias{corn}
\title{Data of corn}
\format{
A data frame with 34 observations on the following 3 variables.
\describe{ \item{list("method")}{a numeric vector}
\item{list("observation")}{a numeric vector} \item{list("rx")}{a numeric
vector} }
}
\source{
Book: Practical Nonparametric Statistics.
}
\description{
Data from a completely randomized design where four different methods of
growing corn resulted in various yields per acre on various plots of ground
where the four methods were tried. Ordinarily, only one statistical analysis
is used, but here we will use the kuskal-wallis test so that a rough
comparison may be made with the mediasn test.
}
\details{
The observations are ranked from the smallest, 77, of rank 1 to the largest
101, of rank N=34. Ties values receive the averarge rank.
}
\examples{

data(corn)
str(corn)

}
\references{
Practical Nonparametrics Statistics. W.J. Conover. Third
Edition, 1999.
}
\keyword{datasets}
