% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl.R
\name{correl}
\alias{correl}
\title{Correlation Coefficient}
\usage{
correl(x, y, method = "pearson", alternative = "two.sided")
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}

\item{method}{"pearson", "kendall", "spearman"}

\item{alternative}{"two.sided", "less", "greater"}
}
\value{
The correlation of x,y vector with the statistical value and its
probability
}
\description{
An exact correlation for ties or without ties.  Methods of Kendall, Spearman
and Pearson.
}
\examples{

library(agricolae)
data(soil)
with(soil,correl(pH,clay,method="kendall"))
with(soil,correl(pH,clay,method="spearman"))
with(soil,correl(pH,clay,method="pearson"))

}
\references{
Numerical Recipes in C. Second Edition.
}
\seealso{
\code{\link{correlation}}
}
\author{
Felipe de Mendiburu
}
\keyword{multivariate}
