% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotton.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{Data of cotton}
\format{
A \code{data.frame} with 96 observations on the following 5 variables.

 @details
        \itemize{
        \item{\strong{site}} a factor with levels Lima, and Pisco
        \item{\strong{block}} a factor with levels I, II, III, IV, V, and VI
        \item{\strong{lineage}} lineage
        \item{\strong{epoca}} epoca
        \item{\strong{yield}} yield
        }
}
\source{
Book spanish: Metodos estadisticos para la investigacion.  Autor:
Calzada Benza Universidad Nacional Agraria - La Molina - Peru..
}
\usage{
data(cotton)
}
\description{
Data of cotton collected in experiments of two localities in Lima and Pisco,
Peru.
}
\examples{

library(agricolae)
data(cotton)
str(cotton)

}
\references{
Book spanish: Metodos estadisticos para la investigacion.
Autor: Calzada Benza Universidad Nacional Agraria - La Molina - Peru.
}
\keyword{datasets}
