% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.model.R
\name{cv.model}
\alias{cv.model}
\title{Coefficient of the experiment variation}
\usage{
cv.model(x)
}
\arguments{
\item{x}{object of model lm() or AOV()}
}
\value{
Returns the coefficient of variation of the experiment according to
the applied statistical model
}
\description{
It obtains the coefficient of variation of the experiment obtained by models
lm() or aov()
}
\details{
sqrt(MSerror)*100/mean(x)
}
\examples{

# see examples from LSD , Waller-Duncan or HSD and complete with it:
library(agricolae)
# not run
# cv<-cv.model(model)


}
\seealso{
\code{\link{LSD.test}}, \code{\link{HSD.test}},
\code{\link{waller.test}}
}
\author{
Felipe de Mendiburu
}
\keyword{univar}
