% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.similarity.R
\name{cv.similarity}
\alias{cv.similarity}
\title{Coefficient of the similarity matrix variation}
\usage{
cv.similarity(A)
}
\arguments{
\item{A}{matrix of binary data}
}
\value{
Returns the coefficient of variation of the similarity model
}
\description{
This process consists of finding the coefficient of the distances of
similarity of binary tables (1 and 0) as used for scoring molecular marker
data for presence and absence of PCR amplification products.
}
\examples{

# molecular markers.
library(agricolae)
data(markers)
cv<-cv.similarity(markers)


}
\seealso{
\code{\link{similarity}}, \code{\link{resampling.cv} }
}
\author{
Felipe de Mendiburu
}
\keyword{multivariate}
