% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.aug.alpha.R
\name{design.aug.alpha}
\alias{design.aug.alpha}
\title{Augmented alpha design type (0,1)}
\usage{
design.aug.alpha(
  trt,
  ck,
  k,
  r,
  serie = 2,
  seed = 0,
  kinds = "Super-Duper",
  randomization = TRUE
)
}
\arguments{
\item{trt}{Treatments}

\item{ck}{Controls}

\item{k}{size block without controls}

\item{r}{Replications}

\item{serie}{number plot, 1: 11,12; 2: 101,102; 3: 1001,1002}

\item{seed}{seed}

\item{kinds}{method for to randomize}

\item{randomization}{TRUE or FALSE - randomize}
}
\value{
\item{parameters}{Design parameters} \item{statistics }{Design
statistics} \item{sketch}{Design sketch} \item{book}{Fieldbook}
}
\description{
Generates an augmented alpha designs starting from the alpha design fixing
under the series formulated by Patterson and Williams. These designs are
generated by the alpha arrangements. They are similar to the lattice designs,
but the tables are rectangular s by k (with s blocks and k<s columns). The
number of treatments should be equal to s*k, all controls are randomly arranged
in each block and all the experimental units r*s*(k+nck) (r replications and
nck controls).
}
\details{
Parameters for the augmented alpha design: I.  r=2, k <= s; II. r=3, s odd,
k <= s; III.r=3, s even, k <= s-1; IV. r=4, s odd but not a multiple of 3,
k<=s r= replications s=number of blocks k=size of block Number of treatment
is equal to k*s
}
\examples{

library(agricolae)
#Example one
trt<-1:30
ck <- 31:32
t <- length(trt)
# size block k
k<-3
# Blocks s
s<-t/k
# replications r
r <- 2
outdesign<- design.aug.alpha(trt,ck,k,r,serie=2)
book<-outdesign$book
plots<-book[,1]
dim(plots)<-c(k,s,r)
for (i in 1:r) print(t(plots[,,i]))
outdesign$sketch
# Example two
trt<-letters[1:12]
ck <- letters[13:14]
t <- length(trt)
k<-3
r<-3
s<-t/k
outdesign<- design.aug.alpha(trt,ck,k,r,serie=2)
book<-outdesign$book
plots<-book[,1]
dim(plots)<-c(k,s,r)
for (i in 1:r) print(t(plots[,,i]))
outdesign$sketch

}
\references{
H.D. Patterson and E.R. Williams. Biometrika (1976) A new class
of resolvable incomplete block designs. printed in Great Britain.  Online:
http://biomet.oxfordjournals.org/cgi/content/abstract/63/1/83
}
\seealso{
\code{\link{design.ab}}, \code{\link{design.alpha}},
\code{\link{design.split}}, \code{\link{design.bib}},
\code{\link{design.crd}}, \code{\link{design.cyclic}},
\code{\link{design.dau}}, \code{\link{design.graeco}},
\code{\link{design.lattice}}, \code{\link{design.lsd}},
\code{\link{design.rcbd}}, \code{\link{design.strip}}
}
\author{
Yuhang Guo
}
\keyword{design}
