% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genxenv.R
\docType{data}
\name{genxenv}
\alias{genxenv}
\title{Data of potato yield in a different environment}
\format{
A \code{data.frame} with 250 observations on the following 3 variables.

 @details
        \itemize{
        \item{\strong{ENV}} Environments
        \item{\strong{GEN}} Genotypes
        \item{\strong{YLD}} Yield
        }
}
\source{
International Potato Center. CIP - Lima Peru.
}
\usage{
data(genxenv)
}
\description{
50 genotypes and 5 environments.
}
\examples{

library(agricolae)
data(genxenv)
str(genxenv)

}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
