% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glycoalkaloids.R
\docType{data}
\name{Glycoalkaloids}
\alias{Glycoalkaloids}
\title{Data Glycoalkaloids}
\format{
A data frame with 25 observations on the following 2 variables.
\describe{ \item{list("HPLC")}{a numeric vector}
\item{list("spectrophotometer")}{a numeric vector} }
}
\source{
International Potato Center. CIP - Lima Peru.
}
\description{
A measurement of the Glycoalkaloids by two methods: HPLC and
spectrophotometer.
}
\examples{

library(agricolae)
data(Glycoalkaloids)
str(Glycoalkaloids)

}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
