% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\docType{data}
\name{growth}
\alias{growth}
\title{Data growth of trees}
\format{
A data frame with 30 observations on the following 3 variables.
\describe{ \item{list("place")}{a factor with levels \code{L1} \code{L2}}
\item{list("slime")}{a numeric vector} \item{list("height")}{a numeric
vector} }
}
\source{
Experimental data (Pucallpa - Peru)
}
\description{
Data growth of pijuayo trees in several localities.
}
\examples{

library(agricolae)
data(growth)
str(growth)

}
\references{
ICRAF lima Peru.
}
\keyword{datasets}
