% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huasahuasi.R
\docType{data}
\name{huasahuasi}
\alias{huasahuasi}
\title{Data: Rainfall thresholds as support for timing fungicide applications in
the control of potato late blight in Peru}
\format{
The format is: List of 2 ( AUDPC, YIELD ) \describe{
\item{list("block")}{a factor with levels \code{I} \code{II} \code{III}}
\item{list("trt")}{a factor with levels \code{40mm} \code{7-days}
\code{Non-application}} \item{list("clon")}{a factor with levels
\code{C386209.10} \code{C387164.4} \code{Cruza148} \code{Musuq}
\code{Yungay}} \item{list("y1da")}{a numeric vector, Kgr./plot}
\item{list("y2da")}{a numeric vector, Kgr./plot} \item{list("y3ra")}{a
numeric vector, Kgr./plot} \item{list("d44")}{a numeric vector, 44 days}
\item{list("d51")}{a numeric vector, 51 days} \item{list("d100")}{a numeric
vector, 100 days} }
}
\source{
Experimental field, 2003. Data Kindly provided by Wilmer Perez.
}
\description{
Timing fungicide sprays based on accumulated rainfall thresholds can be a
successful component of integrated management packages that include
cultivars with moderate or high levels of resistance to late blight. The
simplicity of measuring accumulated rainfall means that the technology can
potentially be used by resource-poor farmers in developing countries.
}
\details{
The experimental unit was formed by 4 furrows of 1.8 m of length, with
distance between furrows from 0.90 m and between plants of 0.30 m. In each
furrow was installed 5 plants. The experiment had 3 repetitions.  From the
beginning of the experiment were fulfilled the following treatments
Thresholds 40 mm: Apply the fungicide when 40 precipitation mm accumulates.
The minimum interval between applications will be of 7 days.  Schedule 7
days: The applications should be carried out every 7 days calendar.  Without
application: No fungicide application will be made.  The evaluation of the
severity of the late blight in each treatment started to emergency 80
percentage and then evaluations were made every 7 days until being observed
a physiological maturation of the crop.
}
\examples{

library(agricolae)
data(huasahuasi)
names(huasahuasi)
str(huasahuasi$AUDPC)
str(huasahuasi$YIELD)


}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
