% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.freq.R
\name{intervals.freq}
\alias{intervals.freq}
\title{Class intervals}
\usage{
intervals.freq(x)
}
\arguments{
\item{x}{class graph.freq, histogram or numeric}
}
\value{
It show interval classes.
}
\description{
List class intervals.
}
\examples{

library(agricolae)
# example 1
data(growth)
h<-hist(growth$height,plot=FALSE)
intervals.freq(h)
# example 2
x<-seq(10,40,5)
y<-c(2,6,8,7,3,4)
intervals.freq(x)
histogram <- graph.freq(x,counts=y)

}
\seealso{
\code{\link{polygon.freq}}, \code{\link{table.freq}},
\code{\link{stat.freq}}, \code{\link{graph.freq}},
\code{\link{sturges.freq}}, \code{\link{join.freq}},
\code{\link{ogive.freq}}, \code{\link{normal.freq} }
}
\author{
Felipe de Mendiburu
}
\keyword{univar}
