% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall.R
\name{kendall}
\alias{kendall}
\title{Correlation of Kendall}
\usage{
kendall(data1, data2)
}
\arguments{
\item{data1}{vector}

\item{data2}{vector}
}
\value{
The correlation of data1, data2 vector with the statistical value
and its probability
}
\description{
Correlation of Kendall two set. Compute exact p-value with ties.
}
\examples{

library(agricolae)
x <-c(1,1,1,4,2,2,3,1,3,2,1,1,2,3,2,1,1,2,1,2)
y <-c(1,1,2,3,4,4,2,1,2,3,1,1,3,4,2,1,1,3,1,2)
kendall(x,y)

}
\references{
Numerical Recipes in C. Second Edition. Pag 634
}
\seealso{
\code{\link{correlation}}
}
\author{
Felipe de Mendiburu
}
\keyword{nonparametric}
