% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Finding the Kurtosis coefficient}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\item{x }{ The kurtosis of x }
}
\description{
It obtains the value of the kurtosis for a normally distributed variable.
The result is similar to SAS.
}
\examples{

library(agricolae)
x<-c(3,4,5,2,3,4,5,6,4,NA,7)
kurtosis(x)
# value is -0.1517996

}
\seealso{
\code{\link{skewness} }
}
\keyword{univar}
