% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastC.R
\name{lastC}
\alias{lastC}
\title{Setting the last character of a chain}
\usage{
lastC(x)
}
\arguments{
\item{x}{letters}
}
\value{
\item{x }{Returns the last character of a string}
}
\description{
A special function for the group of treatments in the multiple comparison
tests.  Use plot.group.
}
\examples{

library(agricolae)
x<-c("a","ab","b","c","cd")
lastC(x)
# "a" "b" "b" "c" "d"


}
\seealso{
\code{\link{plot.group} }
}
\author{
Felipe de Mendiburu
}
\keyword{manip}
