% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LxT.R
\docType{data}
\name{LxT}
\alias{LxT}
\title{Data Line by tester}
\format{
A data frame with 92 observations on the following 4 variables.
\describe{ \item{replication}{a numeric vector} \item{line}{a numeric
vector} \item{tester}{a numeric vector} \item{yield}{a numeric vector}}
}
\source{
Biometrical Methods in Quantitative Genetic Analysis, Singh,
Chaudhary. 1979
}
\description{
Data frame with yield by line x tester.
}
\keyword{datasets}
