% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\docType{data}
\name{markers}
\alias{markers}
\title{Data of molecular markers}
\format{
A \code{data.frame} with 23 observations on the following 27 variables.

 @details
        \itemize{
        \item{\strong{marker1}} marker1
        \item{\strong{marker2}} marker2
        \item{\strong{marker3}} marker3
        \item{\strong{marker4}} marker4
        \item{\strong{marker5}} marker5
        \item{\strong{marker6}} marker6
        \item{\strong{marker7}} marker7
        \item{\strong{marker8}} marker8
        \item{\strong{marker9}} marker9
        \item{\strong{marker10}} marker10
        \item{\strong{marker11}} marker11
        \item{\strong{marker12}} marker12
        \item{\strong{marker13}} marker13
        \item{\strong{marker14}} marker14
        \item{\strong{marker15}} marker15
        \item{\strong{marker16}} marker16
        \item{\strong{marker17}} marker17
        \item{\strong{marker18}} marker18
        \item{\strong{marker19}} marker19
        \item{\strong{marker20}} marker20
        \item{\strong{marker21}} marker21
        \item{\strong{marker22}} marker22
        \item{\strong{marker23}} marker23
        \item{\strong{marker24}} marker24
        \item{\strong{marker25}} marker25
        \item{\strong{marker26}} marker26
        \item{\strong{marker27}} marker27
        }
}
\source{
International Potato Center Lima-Peru.
}
\usage{
data(markers)
}
\description{
A partial study on 27 molecular markers.
}
\examples{

library(agricolae)
data(markers)
str(markers)

}
\references{
International Potato Center Lima-Peru.
}
\keyword{datasets}
