% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natives.R
\docType{data}
\name{natives}
\alias{natives}
\title{Data of native potato}
\format{
A \code{data.frame} with 876 observations on the following 4 variables.

 @details
        \itemize{
        \item{\strong{variety}} variety
        \item{\strong{number}} number
        \item{\strong{weight}} weight
        \item{\strong{size}} size
        }
}
\source{
International Potato Center. CIP - Lima Peru.
}
\usage{
data(natives)
}
\description{
An evaluation of the number, weight and size of 24 native potatoes
varieties.
}
\examples{

library(agricolae)
data(natives)
str(natives)

}
\keyword{datasets}
