% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamCIP.R
\docType{data}
\name{pamCIP}
\alias{pamCIP}
\title{Data Potato Wild}
\format{
A data frame with 43 observations on the following 107 variables.
Rownames: code and genotype's name.  column data: molecular markers.
}
\source{
Laboratory data.
}
\description{
Potato Wild
}
\details{
To study the molecular markers in Wild.
}
\examples{

library(agricolae)
data(pamCIP)
str(pamCIP)

}
\references{
International Potato Center Lima-Peru (CIP)
}
\keyword{datasets}
