% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paracsho.R
\docType{data}
\name{paracsho}
\alias{paracsho}
\title{Data of Paracsho biodiversity}
\format{
A data frame with 110 observations on the following 6 variables.
\describe{ \item{list("date")}{a factor with levels \code{15-12-05}
\code{17-11-05} \code{18-10-05} \code{20-09-05} \code{22-06-05}
\code{23-08-05} \code{28-07-05}} \item{list("plot")}{a factor with levels
\code{PARACSHO}} \item{list("Treatment")}{a factor with levels \code{CON}
\code{SIN}} \item{list("Orden")}{a factor with levels \code{COLEOPTERA}
\code{DIPTERA} \code{HEMIPTERA} \code{HYMENOPTERA} \code{LEPIDOPTERA}
\code{NEUROPTERA} \code{NEUROPTERO} \code{NOCTUIDAE}}
\item{list("Family")}{a factor with levels \code{AGROMYZIDAE}
\code{ANTHOCORIDAE} \code{ANTHOMYIIDAE} \code{ANTHOMYLIDAE}
\code{BLEPHAROCERIDAE} \code{BRACONIDAE} \code{BROCONIDAE}
\code{CALUPHORIDAE} \code{CECIDOMYIDAE} \code{CHENEUMONIDAE}
\code{CHNEUMONIDAE} \code{CHRYOMELIDAE} \code{CICADELLIDAE} \code{CULICIDAE}
\code{ERIOCPAMIDAE} \code{HEMEROBIIDAE} \code{ICHNEUMONIDAE}
\code{LOUCHAPIDAE} \code{MIRIDAE} \code{MUSCIDAE} \code{MUSICADAE}
\code{MUSLIDAE} \code{MYCETOPHILIDAE} \code{MYCETOPHILIIDAE}
\code{NENPHALIDAE} \code{NOCLUIDAE} \code{NOCTERIDAE} \code{NOCTUIDAE}
\code{PERALIDAE} \code{PIPUNCULIDAE} \code{PROCTOTRUPIDAE} \code{PSYLLIDAE}
\code{PYRALIDAE} \code{SARCOPHAGIDAE} \code{SARCOPILAGIDAE}
\code{SCATOPHAGIDAE} \code{SCATOPHOGIDAE} \code{SCIARIDAE} \code{SERSIDAE}
\code{SYRPHIDAE} \code{TACHINIDAE} \code{TIPULIDAE}}
\item{list("Number.of.specimens")}{a numeric vector} }
}
\source{
Entomology dataset.
}
\description{
A locality in Peru. A biodiversity.
}
\details{
Country Peru, Deparment Junin, province Tarma, locality Huasahuasi.
}
\examples{

library(agricolae)
data(paracsho)
str(paracsho)

}
\references{
International Potato Center.
}
\keyword{datasets}
