% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato.R
\docType{data}
\name{potato}
\alias{potato}
\title{Data of cutting}
\format{
A \code{data.frame} with 18 observations on the following 4 variables.

 @details
        \itemize{
        \item{\strong{date}} date
        \item{\strong{variety}} a factor with levels Canchan, and Unica
        \item{\strong{harvest}} harvest
        \item{\strong{cutting}} cutting
        }
}
\source{
Experimental data.
}
\usage{
data(potato)
}
\description{
A study on the yield of two potato varieties performed at the CIP
experimental station.
}
\examples{

library(agricolae)
data(potato)
str(potato)

}
\references{
International Potato Center
}
\keyword{datasets}
