% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\docType{data}
\name{rice}
\alias{rice}
\title{Data of Grain yield of rice variety IR8}
\format{
A \code{data.frame} with 36 observations on the following 18 variables.

 @details
        \itemize{
        \item{\strong{V1}} V1
        \item{\strong{V2}} V2
        \item{\strong{V3}} V3
        \item{\strong{V4}} V4
        \item{\strong{V5}} V5
        \item{\strong{V6}} V6
        \item{\strong{V7}} V7
        \item{\strong{V8}} V8
        \item{\strong{V9}} V9
        \item{\strong{V10}} V10
        \item{\strong{V11}} V11
        \item{\strong{V12}} V12
        \item{\strong{V13}} V13
        \item{\strong{V14}} V14
        \item{\strong{V15}} V15
        \item{\strong{V16}} V16
        \item{\strong{V17}} V17
        \item{\strong{V18}} V18
        }
}
\source{
Statistical Procedures for Agriculture Research. Second Edition.
Kwanchai A. Gomez and Arturo A. Gomez. 1976. USA Pag. 481.
}
\usage{
data(rice)
}
\description{
The data correspond to the yield of rice variety IR8 (g/m2) for land
uniformity studies. The growing area is 18x36 meters.

Table 12.1 Measuring Soil Heterogeneity
}
\examples{

library(agricolae)
data(rice)
str(rice)

}
\references{
Statistical Procedures for Agriculture Research. Second Edition.
Kwanchai A. Gomez and Arturo A. Gomez. 1976. USA
}
\keyword{datasets}
