% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Matrix of similarity in binary data}
\usage{
similarity(A)
}
\arguments{
\item{A}{Matrix, data binary}
}
\value{
Distance matrix. Class = dist.
}
\description{
It finds the similarity matrix of binary tables (1 and 0).
}
\examples{

#example table of molecular markers
library(agricolae)
data(markers)
distance<-similarity(markers)
#startgraph
tree<-hclust(distance,method="mcquitty")
plot(tree,col="blue")
#endgraph

}
\seealso{
\code{\link{cv.similarity}}, \code{\link{resampling.cv} }
}
\author{
Felipe de Mendiburu
}
\keyword{models}
