% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinRepAmmi.R
\docType{data}
\name{sinRepAmmi}
\alias{sinRepAmmi}
\title{Data for AMMI without repetition}
\format{
A data frame with 250 observations on the following 3 variables.
\describe{ \item{list("ENV")}{a factor with levels \code{A1} \code{A2}
\code{A3} \code{A4} \code{A5}} \item{list("GEN")}{a numeric vector}
\item{list("YLD")}{a numeric vector} }
}
\source{
Experimental data.
}
\description{
Data frame for AMMI analysis with 50 genotypes in 5 environments.
}
\examples{

library(agricolae)
data(sinRepAmmi)
str(sinRepAmmi)

}
\references{
International Potato Center - Lima Peru.
}
\keyword{datasets}
