% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Finding the skewness coefficient}
\usage{
skewness(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
The skewness of x.
}
\description{
It returns the skewness of a distribution. It is similar to SAS.
}
\examples{

library(agricolae)
x<-c(3,4,5,2,3,4,NA,5,6,4,7)
skewness(x)
# value is 0,3595431, is slightly asimetrica (positive) to the right

}
\seealso{
\code{\link{kurtosis} }
}
\keyword{univar}
