% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.plot.R
\name{sp.plot}
\alias{sp.plot}
\title{Splip-Plot analysis}
\usage{
sp.plot(block, pplot, splot, Y)
}
\arguments{
\item{block}{replications}

\item{pplot}{main-plot Factor}

\item{splot}{sub-plot Factor}

\item{Y}{Variable, response}
}
\value{
ANOVA: Splip plot analysis
}
\description{
The variance analysis of a split plot design is divided into two parts: the
plot-factor analysis and the sub-plot factor analysis.
}
\details{
The split-plot design is specifically suited for a two-factor experiment on
of the factors is assigned to main plot (main-plot factor), the second
factor, called the subplot factor, is assigned into subplots.
}
\examples{

library(agricolae)
data(plots)
model<-with(plots,sp.plot(block,A,B,yield))
# with aov
plots[,1]<-as.factor(plots[,1])
AOV <- aov(yield ~ block + A*B + Error(block/A),data=plots)
summary(AOV)

}
\references{
Statistical procedures for agricultural research.  Kwanchai A.
Gomez, Arturo A. Gomez. Second Edition. 1984.
}
\seealso{
\code{\link{ssp.plot}}, \code{\link{strip.plot}},
\code{\link{design.split}}, \code{\link{design.strip} }
}
\author{
Felipe de Mendiburu
}
\keyword{models}
