\name{stable}
\alias{stable}
\docType{data}
\title{ Data for stability analysis}
\description{
  Matrix of averages, by rows the genotypes and columns the environmet.
}
\usage{data(stable)}
\format{
  A data frame with 17 observations on the following 12 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a numeric vector}
    \item{\code{V7}}{a numeric vector}
    \item{\code{V8}}{a numeric vector}
    \item{\code{V9}}{a numeric vector}
    \item{\code{V10}}{a numeric vector}
    \item{\code{V11}}{a numeric vector}
    \item{\code{V12}}{a numeric vector}
  }
}

\source{
  CIP. Lima-Peru.
}
\references{
  CIP. Lima-Peru.
}
\examples{
library(agricolae)
data(stable)
str(stable)
}
\keyword{datasets}
