% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.freq.R
\name{stat.freq}
\alias{stat.freq}
\title{Descriptive measures of grouped data}
\usage{
stat.freq(histogram)
}
\arguments{
\item{histogram}{Object create by function hist()}
}
\value{
Statistics of grouped data.
}
\description{
By this process the variance and central measures ar found: average, medium
and mode of grouped data.
}
\examples{

library(agricolae)
data(growth)
grouped<-with(growth,hist(height,plot=FALSE))
measures<-stat.freq(grouped)
print(measures)


}
\seealso{
\code{\link{polygon.freq}}, \code{\link{table.freq}},
\code{\link{graph.freq}}, \code{\link{intervals.freq}},
\code{\link{sturges.freq}}, \code{\link{join.freq}},
\code{\link{ogive.freq}}, \code{\link{normal.freq} }
}
\author{
Felipe de mendiburu
}
\keyword{univar}
