% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vark.R
\name{vark}
\alias{vark}
\title{Variance K, ties, Kendall}
\usage{
vark(x, y)
}
\arguments{
\item{x}{Vector}

\item{y}{vector}
}
\value{
variance of K for Kendall's tau
}
\description{
The Kendall method in order to find the K variance.
}
\details{
Script in C to R.
}
\examples{

library(agricolae)
x <-c(1,1,1,4,2,2,3,1,3,2,1,1,2,3,2,1,1,2,1,2)
y <-c(1,1,2,3,4,4,2,1,2,3,1,1,3,4,2,1,1,3,1,2)
vark(x,y)

}
\references{
Numerical Recipes in C. Second Edition.
}
\seealso{
cor.matrix, cor.vector, cor.mv
}
\author{
Felipe de Mendiburu
}
\keyword{nonparametric}
