% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\docType{data}
\name{disease}
\alias{disease}
\title{Data evaluation of the disease overtime}
\format{
A \code{data.frame} with 21 observations on the following 7 variables.


 @details
        \itemize{
        \item{\strong{plots}} plots
        \item{\strong{rep}} rep
        \item{\strong{trt}} a factor with levels \code{T0}, \code{T1}, \code{T2}, \code{T3}, \code{T4}, \code{T5}, \code{T6}
        \item{\strong{E2}} E2
        \item{\strong{E5}} E5
        \item{\strong{E7}} E7
        \item{\strong{yield}} yield
        }
}
\source{
Experimental data.
}
\usage{
data(disease)
}
\description{
Three evaluations over time and the potato yield when applying several
treatments.
}
\examples{

library(agricolae)
data(disease)
str(disease)

}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
