% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genxenv.R
\docType{data}
\name{genxenv}
\alias{genxenv}
\title{Data of potato yield in a different environment}
\format{
A data frame with 250 observations on the following 3 variables.
\describe{ \item{list("ENV")}{a numeric vector} \item{list("GEN")}{a numeric
vector} \item{list("YLD")}{a numeric vector} }
}
\source{
International Potato Center. CIP - Lima Peru.
}
\description{
50 genotypes and 5 environments.
}
\examples{

library(agricolae)
data(genxenv)
str(genxenv)

}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
