% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glycoalkaloids.R
\docType{data}
\name{Glycoalkaloids}
\alias{Glycoalkaloids}
\title{Data Glycoalkaloids}
\format{
A data frame with 25 observations on the following 2 variables.
}
\source{
International Potato Center. CIP - Lima Peru.
}
\usage{
data(Glycoalkaloids)
}
\description{
A measurement of the Glycoalkaloids by two methods: HPLC and
spectrophotometer.
}
\details{
\itemize{
       \item{\strong{HPLC}} HPLC
       \item{\strong{spectrophotometer}} spectrophotometer
       }
}
\examples{

library(agricolae)
data(Glycoalkaloids)
str(Glycoalkaloids)

}
\references{
International Potato Center. CIP - Lima Peru.
}
\keyword{datasets}
