% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\docType{data}
\name{growth}
\alias{growth}
\title{Data growth of trees}
\format{
A \code{data.frame} with 30 observations on the following 3 variables.

 @details
        \itemize{
        \item{\strong{place}} a factor with levels \code{L1}, and \code{L2}
        \item{\strong{slime}} slime
        \item{\strong{height}} height
        }
}
\source{
Experimental data (Pucallpa - Peru)
}
\usage{
data(growth)
}
\description{
Data growth of pijuayo trees in several localities.
}
\examples{

library(agricolae)
data(growth)
str(growth)

}
\references{
ICRAF lima Peru.
}
\keyword{datasets}
