% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LxT.R
\docType{data}
\name{LxT}
\alias{LxT}
\title{Data Line by tester}
\format{
A \code{data.frame} with 92 observations on the following 4 variables.
\describe{ \item{replication}{a numeric vector} \item{line}{a numeric
vector} \item{tester}{a numeric vector} \item{yield}{a numeric vector}}
}
\usage{
data(LxT)
}
\description{
Data frame with yield by line x tester.
}
\references{
\enumerate{
   \item Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
             }
}
\keyword{datasets}
