\name{LxT}
\alias{LxT}
\docType{data}
\title{ Data Line by tester}
\description{
  Data frame with yield by line x tester.
}
\usage{data(LxT)}
\format{
  A data frame with 92 observations on the following 4 variables.
  \describe{
    \item{replication}{a numeric vector}
    \item{line}{a numeric vector}
    \item{tester}{a numeric vector}
    \item{yield}{a numeric vector}}
}

\source{
  Biometrical Methods in Quantitative Genetic Analysis, Singh, Chaudhary. 1979
}

\keyword{datasets}
