% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\docType{data}
\name{markers}
\alias{markers}
\title{Data of molecular markers}
\format{
A data frame with 23 observations on the following 27 variables.
\describe{ \item{list("marker1")}{a numeric vector} \item{list("marker2")}{a
numeric vector} \item{list("marker3")}{a numeric vector}
\item{list("marker4")}{a numeric vector} \item{list("marker5")}{a numeric
vector} \item{list("marker6")}{a numeric vector} \item{list("marker7")}{a
numeric vector} \item{list("marker8")}{a numeric vector}
\item{list("marker9")}{a numeric vector} \item{list("marker10")}{a numeric
vector} \item{list("marker11")}{a numeric vector} \item{list("marker12")}{a
numeric vector} \item{list("marker13")}{a numeric vector}
\item{list("marker14")}{a numeric vector} \item{list("marker15")}{a numeric
vector} \item{list("marker16")}{a numeric vector} \item{list("marker17")}{a
numeric vector} \item{list("marker18")}{a numeric vector}
\item{list("marker19")}{a numeric vector} \item{list("marker20")}{a numeric
vector} \item{list("marker21")}{a numeric vector} \item{list("marker22")}{a
numeric vector} \item{list("marker23")}{a numeric vector}
\item{list("marker24")}{a numeric vector} \item{list("marker25")}{a numeric
vector} \item{list("marker26")}{a numeric vector} \item{list("marker27")}{a
numeric vector} }
}
\source{
International Potato Center Lima-Peru.
}
\description{
A partial study on 27 molecular markers.
}
\examples{

library(agricolae)
data(markers)
str(markers)

}
\references{
International Potato Center Lima-Peru.
}
\keyword{datasets}
