\name{markers}
\alias{markers}
\docType{data}
\title{ Data of molecular markers}
\description{
  A partial study on 27 molecular markers.
}
\usage{data(markers)}
\format{
  A data frame with 23 observations on the following 27 variables.
  \describe{
    \item{\code{marker1}}{a numeric vector}
    \item{\code{marker2}}{a numeric vector}
    \item{\code{marker3}}{a numeric vector}
    \item{\code{marker4}}{a numeric vector}
    \item{\code{marker5}}{a numeric vector}
    \item{\code{marker6}}{a numeric vector}
    \item{\code{marker7}}{a numeric vector}
    \item{\code{marker8}}{a numeric vector}
    \item{\code{marker9}}{a numeric vector}
    \item{\code{marker10}}{a numeric vector}
    \item{\code{marker11}}{a numeric vector}
    \item{\code{marker12}}{a numeric vector}
    \item{\code{marker13}}{a numeric vector}
    \item{\code{marker14}}{a numeric vector}
    \item{\code{marker15}}{a numeric vector}
    \item{\code{marker16}}{a numeric vector}
    \item{\code{marker17}}{a numeric vector}
    \item{\code{marker18}}{a numeric vector}
    \item{\code{marker19}}{a numeric vector}
    \item{\code{marker20}}{a numeric vector}
    \item{\code{marker21}}{a numeric vector}
    \item{\code{marker22}}{a numeric vector}
    \item{\code{marker23}}{a numeric vector}
    \item{\code{marker24}}{a numeric vector}
    \item{\code{marker25}}{a numeric vector}
    \item{\code{marker26}}{a numeric vector}
    \item{\code{marker27}}{a numeric vector}
  }
}

\source{
  International Potato Center Lima-Peru.
}
\references{
  International Potato Center Lima-Peru.
}
\examples{
library(agricolae)
data(markers)
str(markers)
}
\keyword{datasets}

