% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natives.R
\docType{data}
\name{natives}
\alias{natives}
\title{Data of native potato}
\format{
A data frame with 876 observations on the following 4 variables.
\describe{ \item{list("variety")}{a numeric vector} \item{list("number")}{a
numeric vector} \item{list("weight")}{a numeric vector}
\item{list("size")}{a numeric vector} }
}
\source{
International Potato Center. CIP - Lima Peru.
}
\description{
An evaluation of the number, weight and size of 24 native potatoes
varieties.
}
\examples{

library(agricolae)
data(natives)
str(natives)

}
\keyword{datasets}
