\name{pamCIP}
\alias{pamCIP}
\docType{data}
\title{ Data Potato Wild}
\description{
  Potato Wild
}
\usage{data(pamCIP)}
\format{
  A data frame with 43 observations on the following 107 variables.
  Rownames: code and genotype's name. 
  column data: molecular markers.
}
\details{To study the molecular markers in Wild.
}
\source{
  Laboratory data.
}
\references{
  International Potato Center Lima-Peru (CIP)
}
\examples{
library(agricolae)
data(pamCIP)
str(pamCIP)
}
\keyword{datasets}
