% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato.R
\docType{data}
\name{potato}
\alias{potato}
\title{Data of cutting}
\format{
A data frame with 18 observations on the following 4 variables.
\describe{ \item{list("date")}{a numeric vector} \item{list("variety")}{a
factor with levels \code{Canchan} \code{Unica}} \item{list("harvest")}{a
numeric vector} \item{list("cutting")}{a numeric vector} }
}
\source{
Experimental data.
}
\description{
A study on the yield of two potato varieties performed at the CIP
experimental station.
}
\examples{

library(agricolae)
data(potato)
str(potato)

}
\references{
International Potato Center
}
\keyword{datasets}
