% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\docType{data}
\name{rice}
\alias{rice}
\title{Data of Grain yield of rice variety IR8}
\format{
A data frame with 36 observations on the following 18 variables.
\describe{ \item{list("V1")}{a numeric vector} \item{list("V2")}{a numeric
vector} \item{list("V3")}{a numeric vector} \item{list("V4")}{a numeric
vector} \item{list("V5")}{a numeric vector} \item{list("V6")}{a numeric
vector} \item{list("V7")}{a numeric vector} \item{list("V8")}{a numeric
vector} \item{list("V9")}{a numeric vector} \item{list("V10")}{a numeric
vector} \item{list("V11")}{a numeric vector} \item{list("V12")}{a numeric
vector} \item{list("V13")}{a numeric vector} \item{list("V14")}{a numeric
vector} \item{list("V15")}{a numeric vector} \item{list("V16")}{a numeric
vector} \item{list("V17")}{a numeric vector} \item{list("V18")}{a numeric
vector} }
}
\source{
Statistical Procedures for Agriculture Research. Second Edition.
Kwanchai A. Gomez and Arturo A. Gomez. 1976. USA Pag. 481.
}
\description{
The data correspond to the yield of rice variety IR8 (g/m2) for land
uniformity studies. The growing area is 18x36 meters.
}
\details{
Table 12.1 Measuring Soil Heterogeneity
}
\examples{

library(agricolae)
data(rice)
str(rice)

}
\references{
Statistical Procedures for Agriculture Research. Second Edition.
Kwanchai A. Gomez and Arturo A. Gomez. 1976. USA
}
\keyword{datasets}
