% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinRepAmmi.R
\docType{data}
\name{sinRepAmmi}
\alias{sinRepAmmi}
\title{Data for AMMI without repetition}
\format{
A \code{data.frame} with 250 observations on the following 3 variables.

 @details
        \itemize{
        \item{\strong{ENV}} a factor with levels A1, A2, A3, A4, and A5
        \item{\strong{GEN}} Genotype
        \item{\strong{YLD}} Yield
        }
}
\source{
Experimental data.
}
\usage{
data(sinRepAmmi)
}
\description{
Data frame for AMMI analysis with 50 genotypes in 5 environments.
}
\examples{

library(agricolae)
data(sinRepAmmi)
str(sinRepAmmi)

}
\references{
International Potato Center - Lima Peru.
}
\keyword{datasets}
